requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.42.0
django-ckeditor==6.0.0
django-cors-headers==3.4.0
django-js-asset==1.2.2
djangorestframework==3.8.2
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.2.0
wrapt==1.12.1
